kpmg_assignment/part_1/requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain>=0.1.0

# Azure dependencies
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web and HTTP
requests>=2.31.0

# JSON and data validation
jsonschema>=4.19.0
pydantic>=2.0.0

# Logging and utilities
python-json-logger>=2.0.0

# Image processing
Pillow>=10.0.0

# Additional utilities
typing-extensions>=4.7.0